Remove the `--network-plugin=cni` kublet option.
Remove the `--network-plugin=cni` kublet option. As per the (documentation)[https://kubernetes.io/docs/reference/command-line-tools-reference/kubelet/], this is Docker-specific option and has no effect for the remote/containerd runtime that is being used here. This may be misleading for people that want to go further beyond this guide and modify their clusters.